requirements.txt
psycopg2
django
django-rest-framework
slack_sdk
aiohttp